Issues #1, #2, #3 & #4

+++ src/Controller/TaskController.php
@@ -11,34 +11,60 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 class TaskController extends AbstractController
 {
+    private ManagerRegistry $doctrine;
+
     private TaskService $taskService;
 
-    public function __construct(TaskService $taskService)
+    private TranslatorInterface $translator;
+
+    public function __construct(ManagerRegistry $doctrine, TaskService $taskService, TranslatorInterface $translator)
     {
+        $this->doctrine = $doctrine;
+
         $this->taskService = $taskService;
+
+        $this->translator = $translator;
     }
 
     #[Route('/tasks', name: 'app_task_list', methods: ['GET'])]
-    public function listAction(ManagerRegistry $doctrine): Response
+    #[Route('/tasks/completed', name: 'app_task_completed_list', methods: ['GET'])]
+    public function listAction(Request $request): Response
     {
-        return $this->render('task/list.html.twig', ['tasks' => $doctrine->getRepository(Task::class)->findAll()]);
+        $userId = $this->getUser()->getId();
+
+        $routeName = $request->attributes->get('_route');
+        if ($routeName == 'app_task_list') {
+            $taskIsDone = false;
+        } else {
+            $taskIsDone = true;
+        }
+
+        return $this->render('task/list.html.twig', [
+            'tasks' => $this->doctrine->getRepository(Task::class)->findAll(),
+            'userId' => $userId,
+            'taskIsDone' => $taskIsDone,
+        ]);
     }
 
     #[Route('/tasks/create', name: 'app_task_create', methods: ['GET', 'POST'])]
     public function createAction(Request $request): RedirectResponse|Response
     {
+        $user = $this->getUser();
+
         $task = new Task();
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $this->taskService->addTask($task);
+            $this->taskService->addTask($task, $user);
 
-            $this->addFlash('success', 'La tâche a bien été ajoutée.');
+            $message = $this->translator->trans('The task has been successfully added.');
+            $this->addFlash('success', $message);
 
             return $this->redirectToRoute('app_task_list');
         }
@@ -47,16 +73,24 @@ class TaskController extends AbstractController
     }
 
     #[Route('/tasks/{id}/edit', name: 'app_task_edit', methods: ['GET', 'POST'])]
-    public function editAction(Task $task, Request $request): RedirectResponse|Response
+    public function editAction(Request $request, int $id): RedirectResponse|Response
     {
+        $task = $this->doctrine->getRepository(Task::class)->find($id);
+
+        if (!$task) {
+            $message = $this->translator->trans('The task was not found.');
+            $this->addFlash('error', $message);
+
+            return $this->redirectToRoute('app_task_list');
+        }
+
         $form = $this->createForm(TaskType::class, $task);
-
         $form->handleRequest($request);
-
         if ($form->isSubmitted() && $form->isValid()) {
             $this->taskService->editTask($task);
 
-            $this->addFlash('success', 'La tâche a bien été modifiée.');
+            $message = $this->translator->trans('The task has been successfully edited.');
+            $this->addFlash('success', $message);
 
             return $this->redirectToRoute('app_task_list');
         }
@@ -68,27 +102,54 @@ class TaskController extends AbstractController
     }
 
     #[Route('/tasks/{id}/toggle', name: 'app_task_toggle', methods: ['GET'])]
-    public function toggleAction(Task $task): RedirectResponse
+    public function toggleAction(int $id): RedirectResponse
     {
-        $task->toggle(!$task->isDone());
+        $task = $this->doctrine->getRepository(Task::class)->find($id);
+
+        if (!$task) {
+            $message = $this->translator->trans('The task was not found.');
+            $this->addFlash('error', $message);
+
+            return $this->redirectToRoute('app_task_list');
+        }
+
+        $task->toggle(!$task->getIsDone());
 
         $task = $this->taskService->toggleTask($task);
 
-        $message = $task->getIsDone() ? sprintf('La tâche %s a bien été marquée comme terminée.',
-            $task->getTitle()) : sprintf('La tâche %s a bien été marquée comme non terminée.', $task->getTitle());
-
+        $message = $task->getIsDone() ? 'Task %s was successfully marked as completed.'
+            : 'Task %s was successfully marked as incomplete.';
+        $message = $this->translator->trans($message, ['%s' => $task->getTitle()]);
         $this->addFlash('success', $message);
 
-        return $this->redirectToRoute('app_task_list');
+        if ($task->getIsDone()) {
+            $redirection = $this->redirectToRoute('app_task_completed_list');
+        } else {
+            $redirection = $this->redirectToRoute('app_task_list');
+        }
+
+        return $redirection;
     }
 
     #[Route('/tasks/{id}/delete', name: 'app_task_delete', methods: ['GET'])]
-    public function deleteAction(Task $task): RedirectResponse
+    public function deleteAction(Request $request, int $id): RedirectResponse
     {
-        $this->taskService->deleteTask($task);
+        $user = $this->getUser();
 
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
+        $task = $this->doctrine->getRepository(Task::class)->find($id);
 
-        return $this->redirectToRoute('app_task_list');
+        if (!$task) {
+            throw $this->createAccessDeniedException();
+        } elseif (($task->getUser() == $user) || (!$task->getUser() && $this->isGranted('ROLE_ADMIN'))) {
+            $this->taskService->deleteTask($task);
+
+            $message = $this->translator->trans('The task has been successfully deleted.');
+            $this->addFlash('success', $message);
+
+            $route = $request->headers->get('referer');
+            return $this->redirect($route);
+        } else {
+            throw $this->createAccessDeniedException();
+        }
     }
 }
